Rearranged folder project hierarchy
