+++ pokemon-collection/src/layouts/Footer.tsx
