+++ pokemon-collection/src/layouts/Layout.tsx
@@ -2,7 +2,7 @@ import { Outlet } from 'react-router-dom';
 // components
 import Header from './header/Header';
 import Navbar from './Navbar';
-import Footer from './footer/footer';
+import Footer from './Footer';
 
 const Layout = () => {
   return (
