+++ pokemon-collection/src/pages/ProgressTrackerPage.tsx
@@ -1,5 +1,5 @@
-import AddTrackedRomForm from "../tracker/AddTrackedRomForm";
-import type Rom from "../../types/Rom";
+import AddTrackedRomForm from '../components/tracker/AddTrackedRomForm';
+import type Rom from '../types/Rom';
 
 type Props = {
   trackedRoms: Rom[];
@@ -24,8 +24,8 @@ export default function ProgressTrackerPage({
 
     setTrackedRoms((prev) =>
       prev.map((r) =>
-        r.title === title ? { ...r, percentComplete: clamped } : r
-      )
+        r.title === title ? { ...r, percentComplete: clamped } : r,
+      ),
     );
   }
 
@@ -35,7 +35,7 @@ export default function ProgressTrackerPage({
 
       <div style={{ marginBottom: 12 }}>
         <p>
-          You are feeling: <strong>{sharedMessage || "not sure yet"}</strong>
+          You are feeling: <strong>{sharedMessage || 'not sure yet'}</strong>
         </p>
 
         <label>
@@ -52,19 +52,22 @@ export default function ProgressTrackerPage({
 
       <p>Tracked ROMs: {trackedRoms.length}</p>
 
-      <AddTrackedRomForm trackedRoms={trackedRoms} setTrackedRoms={setTrackedRoms} />
+      <AddTrackedRomForm
+        trackedRoms={trackedRoms}
+        setTrackedRoms={setTrackedRoms}
+      />
 
       {trackedRoms.length === 0 ? (
         <p>No ROMs tracked yet.</p>
       ) : (
         <ul>
           {trackedRoms.map((rom) => {
-            const title = rom.title ?? "";
-            const key = rom.title ?? "unknown-title";
+            const title = rom.title ?? '';
+            const key = rom.title ?? 'unknown-title';
 
             return (
               <li key={key} style={{ marginBottom: 12 }}>
-                <strong>{title || "Untitled ROM"}</strong>
+                <strong>{title || 'Untitled ROM'}</strong>
 
                 <div>
                   Progress:
@@ -73,7 +76,9 @@ export default function ProgressTrackerPage({
                     min={0}
                     max={100}
                     value={rom.percentComplete ?? 0}
-                    onChange={(e) => updateProgress(title, Number(e.target.value))}
+                    onChange={(e) =>
+                      updateProgress(title, Number(e.target.value))
+                    }
                     style={{ marginLeft: 8, width: 60 }}
                     disabled={!title}
                   />
